Move fetchSchema and remove mainViewState as its no longer used
