+++ src/view/dataElements/variable/components/xdmVariable.jsx
@@ -17,8 +17,8 @@ import { useField } from "formik";
 import PropTypes from "prop-types";
 import UserReportableError from "../../../errors/userReportableError";
 import fetchSandboxes from "../../../utils/fetchSandboxes";
-import fetchSchema from "../../../components/objectEditor/helpers/fetchSchema";
-import fetchSchemasMeta from "../../../components/objectEditor/helpers/fetchSchemasMeta";
+import fetchSchema from "../../../utils/fetchSchema";
+import fetchSchemasMeta from "../../../utils/fetchSchemasMeta";
 import FormikPicker from "../../../components/formikReactSpectrum3/formikPicker";
 import FormikPagedComboBox from "../../../components/formikReactSpectrum3/formikPagedComboBox";
 import useReportAsyncError from "../../../utils/useReportAsyncError";
