+++ src/view/utils/fetchSchema.js
@@ -10,8 +10,8 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 
-import fetchFromPlatform from "../../../utils/fetchFromPlatform";
-import UserReportableError from "../../../errors/userReportableError";
+import fetchFromPlatform from "./fetchFromPlatform";
+import UserReportableError from "../errors/userReportableError";
 
 export default async ({
   orgId,
